Create NotesView interface
This is the view part of the MVP.
It's methods will be implemented in MainActivity that implements NotesView interaface

setNotes' @param notes need to be taken from model and directly passed to method by a presenter
and this presenter will be called in the related state by the MainActivity (onCreate())

Idea for removeNote() : longClick on a listItem will fire the presenter.onLongClick method
(so the presenter is the one that decides what happens with data) and this will call the
related method it implemented from a usecase interface - or it may directly call the view's method.
When it finished it will call the interface that MainActivity
implemented to update the View when its called.

Presenter uses an extra interface if the method it needs to call takes model as a parameter.
If this is like longClick and the position is already given to presenter by onLongItemClick()
method of ArrayAdapter, then there's no need to use a another interface and presenter will directly
call the view's method which is implemented by activity to update view.